+++ projects/observability/src/shared/dashboard/data/graphql/api-calls-count/api-calls-count-data-source.model.test.ts
@@ -3,11 +3,11 @@ import { AttributeMetadataType, MetricAggregationType, ObservedGraphQlRequest } 
 import { ModelApi } from '@hypertrace/hyperdash';
 import { runFakeRxjs } from '@hypertrace/test-utils';
 import { ExploreSpecificationBuilder } from '../../../../graphql/request/builders/specification/explore/explore-specification-builder';
+import { ExploreGraphQlQueryHandlerService } from '../../../../graphql/request/handlers/explore/explore-graphql-query-handler.service';
 import {
-  ExploreGraphQlQueryHandlerService,
   EXPLORE_GQL_REQUEST,
   GraphQlExploreResponse
-} from '../../../../graphql/request/handlers/explore/explore-graphql-query-handler.service';
+} from '../../../../graphql/request/handlers/explore/explore-query';
 import { ApiCallsCountDataSourceModel } from './api-calls-count-data-source-model';
 
 describe('API call count data source model', () => {
