+++ projects/observability/src/shared/dashboard/data/graphql/metric-aggregation/metric-aggregation-data-source.model.ts
@@ -9,11 +9,11 @@ import {
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { ExploreSpecificationBuilder } from '../../../../graphql/request/builders/specification/explore/explore-specification-builder';
+import { ExploreGraphQlQueryHandlerService } from '../../../../graphql/request/handlers/explore/explore-graphql-query-handler.service';
 import {
-  ExploreGraphQlQueryHandlerService,
   EXPLORE_GQL_REQUEST,
   GraphQlExploreResponse
-} from '../../../../graphql/request/handlers/explore/explore-graphql-query-handler.service';
+} from '../../../../graphql/request/handlers/explore/explore-query';
 import { ExploreSelectionSpecificationModel } from '../specifiers/explore-selection-specification.model';
 
 @Model({
