+++ projects/observability/src/shared/dashboard/data/graphql/trace/trace-value-data-source.model.ts
@@ -3,12 +3,12 @@ import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { ObservabilityTraceType } from '../../../../graphql/model/schema/observability-traces';
 import { ExploreSpecification } from '../../../../graphql/model/schema/specifications/explore-specification';
+import { ExploreGraphQlQueryHandlerService } from '../../../../graphql/request/handlers/explore/explore-graphql-query-handler.service';
 import {
-  ExploreGraphQlQueryHandlerService,
   EXPLORE_GQL_REQUEST,
   GraphQlExploreRequest,
   GraphQlExploreResultValue
-} from '../../../../graphql/request/handlers/explore/explore-graphql-query-handler.service';
+} from '../../../../graphql/request/handlers/explore/explore-query';
 
 export abstract class TraceValueDataSourceModel<TData> extends GraphQlDataSourceModel<TData> {
   protected abstract specification: ExploreSpecification;
