+++ projects/observability/src/shared/graphql/request/handlers/explore/explore-graphql-query-builder.service.ts
@@ -1,20 +1,21 @@
 import { Injectable } from '@angular/core';
-import { DateCoercer, Dictionary, TimeDuration } from '@hypertrace/common';
-import {
-  GlobalGraphQlFilterService,
-  GraphQlFilter,
-  GraphQlSelectionBuilder,
-  GraphQlSortBySpecification,
-  GraphQlTimeRange
-} from '@hypertrace/distributed-tracing';
+import { DateCoercer, Dictionary } from '@hypertrace/common';
+import { GlobalGraphQlFilterService, GraphQlSelectionBuilder } from '@hypertrace/distributed-tracing';
 import { GraphQlArgument, GraphQlEnumArgument, GraphQlSelection } from '@hypertrace/graphql-client';
 import { INTERVAL_START_QUERY_KEY } from '../../../model/schema/explore';
 import { GraphQlGroupBy } from '../../../model/schema/groupby/graphql-group-by';
-import { ExploreSpecification, ExploreValue } from '../../../model/schema/specifications/explore-specification';
+import { ExploreSpecification } from '../../../model/schema/specifications/explore-specification';
 import { GraphQlObservabilityArgumentBuilder } from '../../builders/argument/graphql-observability-argument-builder';
 import { ExploreSpecificationBuilder } from '../../builders/specification/explore/explore-specification-builder';
-
-export const GQL_EXPLORE_RESULT_INTERVAL_KEY = Symbol('Interval Start');
+import {
+  GQL_EXPLORE_RESULT_INTERVAL_KEY,
+  GraphQlExploreRequest,
+  GraphQlExploreResponse,
+  GraphQlExploreResult,
+  GraphQlExploreResultValue,
+  GraphQlExploreServerResponse,
+  GraphQlExploreServerResult
+} from './explore-query';
 
 @Injectable({ providedIn: 'root' })
 export class ExploreGraphqlQueryBuilderService {
@@ -102,46 +103,4 @@ export class ExploreGraphqlQueryBuilderService {
   private groupByAsSpecifications(groupBy?: GraphQlGroupBy): ExploreSpecification[] {
     return (groupBy?.keys ?? []).map(key => this.specBuilder.exploreSpecificationForKey(key));
   }
-}
-
-export const EXPLORE_GQL_REQUEST = Symbol('GraphQL Query Request');
-
-export interface GraphQlExploreRequest {
-  requestType: typeof EXPLORE_GQL_REQUEST;
-  selections: ExploreSpecification[];
-  context: string; // Scope of Request. Example: Trace, Span
-  limit: number;
-  timeRange: GraphQlTimeRange;
-  offset?: number;
-  includeTotal?: boolean;
-  interval?: TimeDuration;
-  orderBy?: GraphQlSortBySpecification[];
-  filters?: GraphQlFilter[];
-  groupBy?: GraphQlGroupBy;
-}
-
-export interface GraphQlExploreResponse {
-  total?: number;
-  results: GraphQlExploreResult[];
-}
-
-export interface GraphQlExploreResult {
-  [GQL_EXPLORE_RESULT_INTERVAL_KEY]?: Date;
-  [key: string]: GraphQlExploreResultValue;
-}
-
-type GraphQlExploreValueType = number | string | boolean | undefined;
-
-export interface GraphQlExploreResultValue extends ExploreValue<GraphQlExploreValueType> {
-  units?: string;
-}
-
-export interface GraphQlExploreServerResponse {
-  total?: number;
-  results: GraphQlExploreServerResult[];
-}
-
-export interface GraphQlExploreServerResult {
-  [INTERVAL_START_QUERY_KEY]?: string;
-  [key: string]: GraphQlExploreResultValue | string | undefined;
 }
